@@ -35,20 +35,20 @@
     "@ionic/angular": ">=7.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.1303.11",
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-devkit/core": "~13.3.11",
-    "@angular-devkit/schematics": "~13.3.11",
-    "@angular/cli": "~13.3.11",
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular-devkit/architect": "^0.1402.11",
+    "@angular-devkit/build-angular": "^14.2.11",
+    "@angular-devkit/core": "^14.2.11",
+    "@angular-devkit/schematics": "^14.2.11",
+    "@angular/cli": "^14.2.11",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@ionic-native/core": "5.0.0-beta.15",
     "@ionic-native/splash-screen": "5.0.0-beta.15",
     "@ionic-native/status-bar": "5.0.0-beta.15",
@@ -78,7 +78,7 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.3.1",
+    "ng-packagr": "^14.2.2",
     "protractor": "~7.0.0",
     "rxjs": "~6.6.0",
     "sw-toolbox": "3.6.0",
