@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { PortService } from '../../services';
 import { Port } from '../../types';
 
@@ -10,12 +10,12 @@ import { Port } from '../../types';
 })
 export class FormControlPage implements OnInit {
   ports: Port[];
-  portControl: FormControl;
-  form: FormGroup;
+  portControl: UntypedFormControl;
+  form: UntypedFormGroup;
 
   constructor(
     private portService: PortService,
-    private formBuilder: FormBuilder
+    private formBuilder: UntypedFormBuilder
   ) { }
 
   ngOnInit() {
